Refactoring to use CustomBinaryExpression.
