+++ src/EFCore.PG/NpgsqlRangeExtensions.cs
@@ -24,7 +24,7 @@
 #endregion
 
 using System;
-using Npgsql.EntityFrameworkCore.PostgreSQL.Query.Expressions.Internal;
+using Npgsql.EntityFrameworkCore.PostgreSQL.Query.ExpressionTranslators.Internal;
 using NpgsqlTypes;
 
 namespace Npgsql.EntityFrameworkCore.PostgreSQL
